+++ 1.21.10/Common/src/main/java/fuzs/universalenchants/UniversalEnchants.java
@@ -3,6 +3,7 @@ package fuzs.universalenchants;
 import fuzs.puzzleslib.api.config.v3.ConfigHolder;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.api.core.v1.context.PackRepositorySourcesContext;
+import fuzs.puzzleslib.api.core.v1.context.PayloadTypesContext;
 import fuzs.puzzleslib.api.core.v1.utility.ResourceLocationHelper;
 import fuzs.puzzleslib.api.event.v1.FinalizeItemComponentsCallback;
 import fuzs.puzzleslib.api.event.v1.core.EventPhase;
@@ -13,6 +14,7 @@ import fuzs.universalenchants.config.ServerConfig;
 import fuzs.universalenchants.handler.BetterEnchantsHandler;
 import fuzs.universalenchants.handler.ItemCompatHandler;
 import fuzs.universalenchants.init.ModRegistry;
+import fuzs.universalenchants.network.ClientboundStopUsingItemMessage;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import org.apache.logging.log4j.LogManager;
@@ -48,6 +50,12 @@ public class UniversalEnchants implements ModConstructor {
         // run after other mods had a chance to change looting level
         LivingExperienceDropCallback.EVENT.register(EventPhase.AFTER, BetterEnchantsHandler::onLivingExperienceDrop);
         BlockEvents.DROP_EXPERIENCE.register(EventPhase.AFTER, BetterEnchantsHandler::onDropExperience);
+    }
+
+    @Override
+    public void onRegisterPayloadTypes(PayloadTypesContext context) {
+        context.optional();
+        context.playToClient(ClientboundStopUsingItemMessage.class, ClientboundStopUsingItemMessage.STREAM_CODEC);
     }
 
     @Override